+++ Pester.BeforeContainer.ps1
@@ -1,0 +1,30 @@
+<#
+    .SYNOPSIS
+        Bootstrap script for Pester tests that runs before each container (each test file)
+
+    .DESCRIPTION
+        Bootstrap script for Pester tests that runs before each container (each test file)
+#>
+Write-Information -MessageData 'Invoking Pester.BeforeContainer.' -InformationAction 'Continue'
+
+try
+{
+    if (-not (Get-Module -Name 'DscResource.Test'))
+    {
+        # Assumes dependencies has been resolved, so if this module is not available, run 'noop' task.
+        if (-not (Get-Module -Name 'DscResource.Test' -ListAvailable))
+        {
+            # Redirect all streams to $null, except the error stream (stream 2)
+            & "$PSScriptRoot/build.ps1" -Tasks 'noop' 3>&1 4>&1 5>&1 6>&1 > $null
+        }
+
+        # If the dependencies has not been resolved, this will throw an error.
+        Import-Module -Name 'DscResource.Test' -Force -ErrorAction 'Stop'
+    }
+}
+catch [System.IO.FileNotFoundException]
+{
+    throw 'DscResource.Test module dependency not found. Please run ".\build.ps1 -ResolveDependency -Tasks build" first.'
+}
+
+Write-Information -MessageData 'Successfully invoked Pester.BeforeContainer.' -InformationAction 'Continue'