+++ tests/QA/module.tests.ps1
@@ -239,4 +239,3 @@ Describe 'Help for module' -Tags 'helpQuality' {
         }
     }
 }
-